Cast pointer types to `uintptr_t`
Adds casts of pointer types to `uintptr_t` instead of fixed sized
integer types. This makes these casts stable across changes of
pointer sizes.

+++ libsel4muslcsys/src/sys_morecore.c
@@ -289,7 +289,7 @@ sys_mremap_dynamic(va_list ap)
 
     assert(flags == MREMAP_MAYMOVE);
     assert(IS_ALIGNED_4K(old_size));
-    assert(IS_ALIGNED_4K((uint32_t) old_address));
+    assert(IS_ALIGNED_4K((uintptr_t) old_address));
     assert(IS_ALIGNED_4K(new_size));
     /* we currently only support remaping to size >= original */
     assert(new_size >= old_size);
